move common files into stream_common folder
