+++ benchmarks/pointer_chase/stream_pointer_reader.c
@@ -1,42 +1,6 @@
-#include <errno.h>
-#include <sys/ioctl.h>
-#include <stdio.h>
-#include <sys/types.h>
-#include <sys/stat.h>
-#include <fcntl.h>
-#include <stdlib.h>
-#include <stdint.h>
-#include <string.h>
-#include <unistd.h>
-#include <errno.h>
-#include <time.h>
-#include <pthread.h>
-#include <linux/nvme_ioctl.h>
-
+#include "../stream_common/common.h"
 #define DEBUG 1
 #include "common.h"
-#define POSIX_FADV_STREAMID 8 
-#define IO_TRANSFER_SIZE (4*1024) 
-#define IO_ST_TRANSFER_SIZE (16*1024) 
-#define IO_SEGMENT_SIZE (IO_TRANSFER_SIZE * 8) 
-#define MAX_FILE_OFFSET (IO_TRANSFER_SIZE * 10)
-
-#define IO_OFFSET_NW 0x000 //0xa000 
-
-#define nvme_admin_directive_send 0x19
-#define nvme_admin_directive_recv 0x1a
-
-//#define IO_OPEN_OPTIONS (O_RDWR | O_DIRECT | O_LARGEFILE)
-#define IO_OPEN_OPTIONS (O_RDWR | O_DIRECT | O_NONBLOCK | O_ASYNC)
-
-static unsigned long next;
-
-struct sdm {
-	char *fn;
-	unsigned int sid;
-	unsigned char *data_in;
-	unsigned char *data_out;
-};
 
 int enable_computational_stream_directive (int fd);
 int disable_computational_stream_directive (int fd);
@@ -102,7 +66,6 @@ void computational_read(void *x)
 			printf("enable computational stream directive successful\n");
 		}
 
-		posix_fadvise(fd, f->sid, 0, POSIX_FADV_STREAMID);
 		// IO_TRANSFER_SIZE = 4K 
 		// Grannularity of Transfer
 		// A read on a pointer_chase_enabled stream should return end of block with END_BLOCK_MAGIC
@@ -150,74 +113,6 @@ void computational_read(void *x)
 	fclose(head_pointer_list_ptr);
 }
 
-int nvme_dir_send(int fd, __u32 nsid, __u16 dspec, __u8 dtype, __u8 doper,
-                  __u32 data_len, __u32 dw12, void *data, __u32 *result)
-{
-        struct nvme_admin_cmd cmd = {
-                .opcode         = nvme_admin_directive_send,
-                .addr           = (__u64)(uintptr_t) data,
-                .data_len       = data_len,
-                .nsid           = nsid,
-                .cdw10          = data_len? (data_len >> 2) - 1 : 0,
-                .cdw11          = dspec << 16 | dtype << 8 | doper,
-                .cdw12          = dw12,
-        };
-        int err;
-
-        err = ioctl(fd, NVME_IOCTL_ADMIN_CMD, &cmd);
-        if (!err && result)
-                *result = cmd.result;
-        return err;
-}
-
-int nvme_dir_recv(int fd, __u32 nsid, __u16 dspec, __u8 dtype, __u8 doper,
-                  __u32 data_len, __u32 dw12, void *data, __u32 *result)
-{
-        struct nvme_admin_cmd cmd = {
-                .opcode         = nvme_admin_directive_recv,
-                .addr           = (__u64)(uintptr_t) data,
-                .data_len       = data_len,
-                .nsid           = nsid,
-                .cdw10          = data_len? (data_len >> 2) - 1 : 0,
-                .cdw11          = dspec << 16 | dtype << 8 | doper,
-                .cdw12          = dw12,
-        };
-        int err;
-
-        err = ioctl(fd, NVME_IOCTL_ADMIN_CMD, &cmd);
-        if (!err && result)
-                *result = cmd.result;
-        return err;
-}
-
-int nvme_get_nsid(int fd)
-{
-        static struct stat nvme_stat;
-        int err = fstat(fd, &nvme_stat);
-
-        if (err < 0)
-                return err;
-
-        if (!S_ISBLK(nvme_stat.st_mode)) {
-                fprintf(stderr,
-                        "Error: requesting namespace-id from non-block device\n");
-                exit(ENOTBLK);
-        }
-        return ioctl(fd, NVME_IOCTL_ID);
-}
-
-int enable_stream_directive(int fd)
-{
-	__u32 result;
-	int err;
-	int nsid = nvme_get_nsid(fd);
-
-	printf("Enable stream directive for nsid %d\n", nsid);
-	// dspec, dtype, doper, length, dw12, data, result
-	err = nvme_dir_send(fd, nsid, 0, 0, 1, 0, 0x101, NULL, &result);
-	return err;
-}
-
 /* 0x401 = 0b 0100 0000 0001 this enables counting (dw12 << 10 = 01)
    0x801 = 0b 1000 0000 0001 this enables pointer chase (dw12 << 10 = 02)
 */
@@ -244,19 +139,6 @@ int disable_computational_stream_directive (int fd)
 	printf("Enable stream directive for nsid %d\n", nsid);
 	// dspec, dtype, doper, length, dw12, data, result
 	err = nvme_dir_send(fd, nsid, 0, 0, 1, 0, 0x800, NULL, &result);
-	return err;
-}
-
-int alloc_stream_resources(int fd, unsigned int rsc_cnt)
-{
-	__u32 result;
-	int err;
-	int nsid = nvme_get_nsid(fd);
-
-	printf("Allocate stream resource for nsid %d\n", nsid);
-	err = nvme_dir_recv(fd, nsid, 0, 1, 3, 0, rsc_cnt, NULL, &result);
-	if (err==0)
-		printf("  requested %d; returned %d\n", rsc_cnt, result & 0xffff);
 	return err;
 }
 
@@ -304,8 +186,6 @@ int main(int argc, char **argv)
 	}else{
 		printf("allocate stream resource successful\n");
 	}
-
-	f.sid = 1;
 
 	computational_read(&f);
